Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -11,7 +11,7 @@ int main() {
         int u =a-b;
         printf("%d",u);
     } else if(c == '*') {
-        float e = a*b;
+        int e = a*b;
         int(d)=(int)e;
         printf("%d",d);
     } else if(c == '/'){
